requirements.txt
paho-mqtt
pycryptodome
pandas
cryptography
numpy
pyyaml
tqdm
matplotlib
seaborn